requirement.txt
tensorflow>=2.8.0
Pillow>=9.0.0
piexif>=1.1.3
numpy>=1.21.0
matplotlib>=3.4.0